Refactor implementations of StructInfo::derive and TypeBuilderAttr
